Feat:优化代码结构,便于后续维护
