+++ src/autocoder_nano/version.py
@@ -1,3 +1,3 @@
-__version__ = "0.1.25"
+__version__ = "0.1.26"
 __author__ = "moofs"
 __license__ = "Apache License 2.0"
